+++ warp_drive/training/pytorch_lightning.py
@@ -30,7 +30,7 @@ from torch.utils.data import DataLoader, Dataset
 
 from warp_drive.training.algorithms.policygradient.a2c import A2C
 from warp_drive.training.algorithms.policygradient.ppo import PPO
-from warp_drive.training.models.fully_connected import FullyConnected
+from warp_drive.training.models.factory import ModelFactory
 from warp_drive.training.trainer import Metrics
 from warp_drive.training.utils.data_loader import create_and_push_data_placeholders
 from warp_drive.training.utils.param_scheduler import LRScheduler, ParamScheduler
@@ -353,17 +353,15 @@ class WarpDriveModule(LightningModule):
 
     def _initialize_policy_model(self, policy):
         policy_model_config = self._get_config(["policy", policy, "model"])
-        if policy_model_config["type"] == "fully_connected":
-            model = FullyConnected(
-                self.cuda_envs,
-                policy_model_config["fc_dims"],
-                policy,
-                self.policy_tag_to_agent_id_map,
-                self.create_separate_placeholders_for_each_policy,
-                self.obs_dim_corresponding_to_num_agents,
-            )
-        else:
-            raise NotImplementedError
+        model_obj = ModelFactory.create(policy_model_config["type"])
+        model = model_obj(
+            env=self.cuda_envs,
+            model_config=policy_model_config,
+            policy=policy,
+            policy_tag_to_agent_id_map=self.policy_tag_to_agent_id_map,
+            create_separate_placeholders_for_each_policy=self.create_separate_placeholders_for_each_policy,
+            obs_dim_corresponding_to_num_agents=self.obs_dim_corresponding_to_num_agents,
+        )
         self.models[policy] = model
 
     def _get_config(self, args):